+++ rm-db-format/src/main/java/org/ehrbase/openehr/dbformat/VersionedObjectDataStructure.java
@@ -17,15 +17,11 @@
  */
 package org.ehrbase.openehr.dbformat;
 
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.databind.node.TextNode;
-import com.nedap.archie.base.OpenEHRBase;
 import com.nedap.archie.rm.RMObject;
 import com.nedap.archie.rm.datavalues.quantity.DvProportion;
 import com.nedap.archie.rm.datavalues.quantity.datetime.DvDate;
@@ -39,7 +35,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
-import org.ehrbase.openehr.sdk.serialisation.jsonencoding.CanonicalJson;
+import org.ehrbase.openehr.dbformat.json.RmDbJson;
 import org.ehrbase.openehr.sdk.util.OpenEHRDateTimeSerializationUtils;
 import org.ehrbase.openehr.sdk.webtemplate.parser.NodeId;
 
@@ -50,23 +46,10 @@ public final class VersionedObjectDataStructure {
      */
     public static final String MAGNITUDE_FIELD = "_magnitude";
 
-    public static final ObjectMapper MARSHAL_OM = CanonicalJson.MARSHAL_OM
-            .copy()
-            .setDefaultTyping(
-                    new ObjectMapper.DefaultTypeResolverBuilder(ObjectMapper.DefaultTyping.EVERYTHING) {
-                        @Override
-                        public boolean useForType(JavaType t) {
-                            return OpenEHRBase.class.isAssignableFrom(t.getRawClass());
-                        }
-                    }.init(JsonTypeInfo.Id.NAME, new CanonicalJson.CJOpenEHRTypeNaming())
-                            .typeProperty(DbToRmFormat.TYPE_ATTRIBUTE)
-                            .typeIdVisibility(true)
-                            .inclusion(JsonTypeInfo.As.PROPERTY));
-
     private VersionedObjectDataStructure() {}
 
     public static List<StructureNode> createDataStructure(RMObject rmObject) {
-        JsonNode jsonNode = MARSHAL_OM.valueToTree(rmObject);
+        JsonNode jsonNode = RmDbJson.MARSHAL_OM.valueToTree(rmObject);
         fillInMagnitudes(jsonNode);
 
         var root = createStructureDto(
@@ -125,19 +108,25 @@ public final class VersionedObjectDataStructure {
                 case "DV_DATE_TIME" -> object.put(
                         MAGNITUDE_FIELD,
                         OpenEHRDateTimeSerializationUtils.toMagnitude(
-                                MARSHAL_OM.treeToValue(object, DvDateTime.class)));
+                                RmDbJson.MARSHAL_OM.treeToValue(object, DvDateTime.class)));
                 case "DV_DATE" -> object.put(
                         MAGNITUDE_FIELD,
-                        OpenEHRDateTimeSerializationUtils.toMagnitude(MARSHAL_OM.treeToValue(object, DvDate.class)));
+                        OpenEHRDateTimeSerializationUtils.toMagnitude(
+                                RmDbJson.MARSHAL_OM.treeToValue(object, DvDate.class)));
                 case "DV_TIME" -> object.put(
                         MAGNITUDE_FIELD,
-                        OpenEHRDateTimeSerializationUtils.toMagnitude(MARSHAL_OM.treeToValue(object, DvTime.class)));
+                        OpenEHRDateTimeSerializationUtils.toMagnitude(
+                                RmDbJson.MARSHAL_OM.treeToValue(object, DvTime.class)));
                 case "DV_DURATION" -> object.put(
                         MAGNITUDE_FIELD,
-                        MARSHAL_OM.treeToValue(object, DvDuration.class).getMagnitude());
+                        RmDbJson.MARSHAL_OM
+                                .treeToValue(object, DvDuration.class)
+                                .getMagnitude());
                 case "DV_PROPORTION" -> object.put(
                         MAGNITUDE_FIELD,
-                        MARSHAL_OM.treeToValue(object, DvProportion.class).getMagnitude());
+                        RmDbJson.MARSHAL_OM
+                                .treeToValue(object, DvProportion.class)
+                                .getMagnitude());
                 default -> {
                     /* do not add magnitude */
                 }
@@ -295,7 +284,7 @@ public final class VersionedObjectDataStructure {
                 String rmNameAlias = RmTypeAlias.getAlias(child.textValue());
                 child = new TextNode(rmNameAlias);
             }
-            newNode.put(alias, child);
+            newNode.putIfAbsent(alias, child);
         });
 
         return newNode;